Initial commit of existing file at tumblr
This file is the current HTML/CSS/Javascript for the briandewitt.org site hosted by Tumblr. 
I removed the Emojli and G+1 links next to the Follow @briandewitt Twitter link. 

Next revision may be to update the list of topics/tags. Not sure. One that needs to be added I think are things that have to do with my child.